// src/api/models/dto_AlbumAddDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { dto_AlbumMediaAddDTO } from './dto_AlbumMediaAddDTO'
export type dto_AlbumAddDTO = {
  cover_img?: string
  description?: string
  id?: number
  medias?: Array<dto_AlbumMediaAddDTO>
  name?: string
  start_time?: string
  user_id?: number
}