// src/api/models/vo_GetPresignedURLVO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type vo_GetPresignedURLVO = {
  /**
   * 过期时间戳
   */
  expireAt?: number
  /**
   * 对象键
   */
  key?: string
  /**
   * 预签名URL
   */
  url?: string
}